+++ sparta.go
@@ -32,7 +32,7 @@ import (
 
 const (
 	// SpartaVersion defines the current Sparta release
-	SpartaVersion = "0.20.4"
+	SpartaVersion = "0.30.0"
 	// NodeJSVersion is the Node JS runtime used for the shim layer
 	NodeJSVersion = "nodejs6.10"
 	// PythonVersion is the Python version used for CGO support
@@ -42,6 +42,15 @@ const (
 	// divider length is the length of a divider in the text
 	// based CLI output
 	dividerLength = 62
+)
+
+const (
+	// spartaEnvVarDiscoveryInformation is the name of the discovery information
+	// published into the environment
+	spartaEnvVarDiscoveryInformation = "SPARTA_DISCOVERY_INFO"
+	// spartaEnvVarBuildID is the environment key that includes the buildID
+	// that this was built with
+	spartaEnvVarBuildID = "SPARTA_BUILD_ID"
 )
 
 var (
@@ -96,13 +105,13 @@ func init() {
 
 // Represents the CloudFormation Arn of this stack, referenced
 // in CommonIAMStatements
-var cloudFormationThisStackArn = []gocf.Stringable{gocf.String("arn:aws:cloudformation:"),
-	gocf.Ref("AWS::Region").String(),
-	gocf.String(":"),
-	gocf.Ref("AWS::AccountId").String(),
-	gocf.String(":stack/"),
-	gocf.Ref("AWS::StackName").String(),
-	gocf.String("/*")}
+// var cloudFormationThisStackArn = []gocf.Stringable{gocf.String("arn:aws:cloudformation:"),
+// 	gocf.Ref("AWS::Region").String(),
+// 	gocf.String(":"),
+// 	gocf.Ref("AWS::AccountId").String(),
+// 	gocf.String(":stack/"),
+// 	gocf.Ref("AWS::StackName").String(),
+// 	gocf.String("/*")}
 
 // CommonIAMStatements defines common IAM::Role Policy Statement values for different AWS
 // service types.  See http://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html#genref-aws-service-namespaces
@@ -133,12 +142,14 @@ var CommonIAMStatements = struct {
 			Effect:   "Allow",
 			Resource: wildcardArn,
 		},
-		{
-			Effect: "Allow",
-			Action: []string{"cloudformation:DescribeStacks",
-				"cloudformation:DescribeStackResource"},
-			Resource: gocf.Join("", cloudFormationThisStackArn...),
-		},
+		// Removed as part of migrating sparta.Discovery
+		// to
+		// {
+		// 	Effect: "Allow",
+		// 	Action: []string{"cloudformation:DescribeStacks",
+		// 		"cloudformation:DescribeStackResource"},
+		// 	Resource: gocf.Join("", cloudFormationThisStackArn...),
+		// },
 		// http://docs.aws.amazon.com/lambda/latest/dg/lambda-x-ray.html#enabling-x-ray
 		{
 			Effect: "Allow",
@@ -417,11 +428,13 @@ type IAMRolePrivilege struct {
 }
 
 func (rolePrivilege *IAMRolePrivilege) resourceExpr() *gocf.StringExpr {
-	switch rolePrivilege.Resource.(type) {
+	switch typedPrivilege := rolePrivilege.Resource.(type) {
 	case string:
-		return gocf.String(rolePrivilege.Resource.(string))
+		return gocf.String(typedPrivilege)
+	case gocf.RefFunc:
+		return typedPrivilege.String()
 	default:
-		return rolePrivilege.Resource.(*gocf.StringExpr)
+		return typedPrivilege.(*gocf.StringExpr)
 	}
 }